refactor: Remove inheritance from constant builder
- Extract interface and replace with 2 implementations one for no
  constants and another that will include all constants
