+++ rector.php
@@ -25,9 +25,6 @@ return static function (ContainerConfigurator $containerConfigurator): void {
 
     $parameters->set(Option::PHP_VERSION_FEATURES, PhpVersion::PHP_72);
 
-    // Run Rector only on changed files
-    $parameters->set(Option::ENABLE_CACHE, true);
-
     // Path to phpstan with extensions, that PHPStan in Rector uses to determine types
     $parameters->set(Option::PHPSTAN_FOR_RECTOR_PATH, getcwd() . '/phpstan.neon');
